Improve ability to import new strings.
We need an easier way to import new strings from a .strings or .xml
file. However, in the normal usecase, we do not want to import
unidentified strings since the strings data file is considered the
master data file and we assume that if it doesn't contain a string that
is found in an input file then that string was recently deleted from the
database.

So, we add a -a option to import all strings we find. We also print
clearer warnings if an unidentified string is found.

Also, clean up some of the code that makes more sense as methods of
StringsFile or StringsRow instead of the abstract formatter class.