Write all languages to the twine strings file.
We used to only write translations to the strings file that did not
match the developer language. It was done this way to try and reduce
space in the strings file and to get en-UK to pick up en changes.

This turned out to be the wrong decision. It's better to explicitly
ensure that all strings are translated in all languages, and to not
automatically pick up en string changes in other languages.

This fixes issue #29.