Added the --alternate-key option
With this option you can specify that Twine has to lookup a different value to use as the key in the output. This is mainly to be used for weird keys that Xcode requires for Storyboards and InfoPlist string files.

For example, Storyboard string files use unwieldy keys like "Lf7-1k-UjD.text". Instead of actually using these as the key names (rendering them incomprehensible for non-iOS projects) you can specifiy "key_storyboard = Lf7-1k-UjD.text" as part of the string definition. This key can then be activated with the "--alternate-key storyboard" option which will use the storyboard key instead of the default key.

If no such key is specified it will default to the normal key.