Remote empty sections before joining them together
Previously, a comma was being printed on the last line of every section, even if
that section did not contribute any rows to the formatted output (e.g., no rows
within that section had the tags needed to be output during this run). When such
a zero-row section happens, it results in an line with only a comma on it, which
is not valid json. This can make incremental adjustment of the jquery files
rather tedious, even when using gui tools to quickly reject changes.